+++ private/bufpkg/bufcheck/client.go
@@ -163,7 +163,6 @@ func (c *client) Lint(
 		return nil
 	}
 	return bufanalysis.NewFileAnnotationSet(
-		nil,
 		annotationsToFileAnnotations(
 			imageToPathToExternalPath(
 				image,
@@ -303,7 +302,6 @@ func (c *client) Breaking(
 		return nil
 	}
 	return bufanalysis.NewFileAnnotationSet(
-		nil,
 		annotationsToFileAnnotations(
 			imageToPathToExternalPath(
 				image,